// server/src/types/api.ts
export type TIncomeLevel =
  | "1"
  | "2"
  | "3"
  | "4"
  | "5"
  | "6"
  | "7"
  | "8"
  | "9"
  | "10";

export type TSidejobCategory =
  | "29231"
  | "29647"
  | "29229"
  | "29228"
  | "29230"
  | "29232"
  | "29233";